[CMPT] Complete Kafka producer
